add tools and improve structure scripts

+++ compute-pro-structure.bash
@@ -8,10 +8,11 @@ TMP=pro.tmp
 
 # END OF CONFIGURATION
 
-[ ! -f $INPUT ] && echo "Input file does not exist" && exit 1
-[ -f $OUTPUT ] && rm -rf $OUTPUT
+[[ ! -f $INPUT ]] && echo -e "\033[0;31mError:\033[0m input file does not exist" && exit 1
+[[ -f $OUTPUT ]] && rm -rf $OUTPUT
+[[ -f $TMP ]] && rm -rf $TMP
 
 predator -a -f$TMP -bconst/stride.dat $INPUT
 python3 utils/process_predator_output.py $TMP > $OUTPUT
 
-rm -rf $TMP
+[[ -f $TMP ]] && rm -rf $TMP
